Switch of delete search index if text search feature flagged off remove     <PackageReference Include="Microsoft.Azure.WebJobs.Extensions.Http" Version="3.0.12" /> as transitively pulled in by DurableTask nuget remove DurableTask from PDF generator (not used)

+++ polaris-pipeline/coordinator/Providers/OrchestrationProvider.cs
@@ -140,12 +140,15 @@ public class OrchestrationProvider : IOrchestrationProvider
         
         try
         {
-            await _searchIndexService.RemoveCaseIndexEntriesAsync(caseId, correlationId);
-            telemetryEvent.RemovedCaseIndexTime = DateTime.UtcNow;
+            if (!_configuration.IsConfigSettingEnabled(FeatureFlags.DisableTextExtractorFeatureFlag))
+            {
+                await _searchIndexService.RemoveCaseIndexEntriesAsync(caseId, correlationId);
+                telemetryEvent.RemovedCaseIndexTime = DateTime.UtcNow;
 
-            await _searchIndexService.WaitForCaseEmptyResultsAsync(caseId, correlationId);
-            telemetryEvent.IndexSettledTime = DateTime.UtcNow;
-            
+                await _searchIndexService.WaitForCaseEmptyResultsAsync(caseId, correlationId);
+                telemetryEvent.IndexSettledTime = DateTime.UtcNow;
+            }
+
             await _blobStorageService.DeleteBlobsByCaseAsync(caseIdAsString, correlationId);
             telemetryEvent.BlobsDeletedTime = DateTime.UtcNow;
